Fix typo in Java Programs name
